Implement fee_discount field, fix several bugs in claim_account_evaluator #2651

+++ libraries/chain/witness_schedule.cpp
@@ -76,7 +76,9 @@ void update_median_witness_props( database& db )
       _wso.account_subsidy_print_rate =
          ( uint64_t( median_account_subsidy_limit ) * STEEM_ACCOUNT_SUBSIDY_PRECISION ) / STEEM_BLOCKS_PER_DAY;
       _wso.single_witness_subsidy_limit =
-         ( median_account_subsidy_limit * STEEM_ACCOUNT_SUBSIDY_PRECISION * STEEM_ACCOUNT_SUBSIDY_BURST_DAYS * STEEM_WITNESS_SUBSIDY_PERCENT ) / STEEM_100_PERCENT;
+         ( uint64_t( median_account_subsidy_limit ) *
+           ( uint64_t( STEEM_ACCOUNT_SUBSIDY_PRECISION ) * uint64_t( STEEM_ACCOUNT_SUBSIDY_BURST_DAYS ) * uint64_t( STEEM_WITNESS_SUBSIDY_PERCENT ) )
+         ) / STEEM_100_PERCENT;
    } );
 
    db.modify( db.get_dynamic_global_properties(), [&]( dynamic_global_property_object& _dgpo )
